Add support for non site wide roles
For non site wide roles, permissions need to be given to individual pages.

Currently, when a user is given a 'non site wide' role within a site,
a page permission is created that grants him permissions to the site's first
page. In the future, the site admin will be able to select the pages for which
the page permissions will be created.